# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Indoor positioning system API for tracking BLE devices",
        "title": "TrackerHub API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@trackerhub.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8022",
    "basePath": "/api",
    "paths": {
        "/api/config/web": {
            "get": {
                "description": "Returns the current web UI configuration including beacon positions and map settings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Get web UI configuration",
                "responses": {
                    "200": {
                        "description": "Web UI configuration",
                        "schema": {
                            "$ref": "#/definitions/trackerHub_backend_internal_models.WebUIConfig"
                        }
                    },
                    "500": {
                        "description": "Failed to load configuration",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Updates the web UI configuration with new beacon positions or map settings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Update web UI configuration",
                "parameters": [
                    {
                        "description": "Web UI configuration",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/trackerHub_backend_internal_models.WebUIConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Configuration updated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to save configuration",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/server-runtime-config": {
            "get": {
                "description": "Returns the current server runtime configuration including MQTT and Kalman filter settings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Get server runtime configuration",
                "responses": {
                    "200": {
                        "description": "Server runtime configuration",
                        "schema": {
                            "$ref": "#/definitions/trackerHub_backend_internal_models.ServerRuntimeConfig"
                        }
                    },
                    "500": {
                        "description": "Failed to load configuration",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Updates the server runtime configuration including MQTT broker settings and Kalman filter parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Update server runtime configuration",
                "parameters": [
                    {
                        "description": "Server runtime configuration",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/trackerHub_backend_internal_models.ServerRuntimeConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Configuration updated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to save configuration",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/trackers": {
            "get": {
                "description": "Returns the current state of all active trackers including their positions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trackers"
                ],
                "summary": "Get all tracker states",
                "responses": {
                    "200": {
                        "description": "Map of tracker states",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "trackerHub_backend_internal_models.KalmanParams": {
            "type": "object",
            "required": [
                "measurementVariance",
                "processVariance"
            ],
            "properties": {
                "measurementVariance": {
                    "description": "Kalman filter measurement variance R",
                    "type": "number",
                    "minimum": 0
                },
                "processVariance": {
                    "description": "Kalman filter process variance Q",
                    "type": "number",
                    "minimum": 0
                }
            }
        },
        "trackerHub_backend_internal_models.MQTTServerConfig": {
            "type": "object",
            "required": [
                "applicationID",
                "brokerHost",
                "brokerPort",
                "topicPattern"
            ],
            "properties": {
                "applicationID": {
                    "description": "e.g., your OrgID",
                    "type": "string"
                },
                "brokerHost": {
                    "type": "string"
                },
                "brokerPort": {
                    "type": "integer"
                },
                "clientID": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean",
                    "default": true
                },
                "live_mqtt_status": {
                    "description": "Live MQTT connection status, not saved to file",
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "topicPattern": {
                    "description": "e.g., \"/device_sensor_data/{ApplicationID}/+/+/+/+\"",
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "trackerHub_backend_internal_models.ServerRuntimeConfig": {
            "type": "object",
            "properties": {
                "kalman": {
                    "$ref": "#/definitions/trackerHub_backend_internal_models.KalmanParams"
                },
                "mqtt": {
                    "$ref": "#/definitions/trackerHub_backend_internal_models.MQTTServerConfig"
                },
                "server": {
                    "$ref": "#/definitions/trackerHub_backend_internal_models.WebServerConfig"
                }
            }
        },
        "trackerHub_backend_internal_models.WebServerConfig": {
            "type": "object",
            "required": [
                "port"
            ],
            "properties": {
                "port": {
                    "description": "Port for the HTTP/WebSocket server",
                    "type": "integer"
                }
            }
        },
        "trackerHub_backend_internal_models.WebUIBeaconConfig": {
            "type": "object",
            "required": [
                "major",
                "minor",
                "txPower",
                "uuid",
                "x",
                "y"
            ],
            "properties": {
                "displayName": {
                    "type": "string"
                },
                "macAddress": {
                    "description": "Physical MAC address if known",
                    "type": "string"
                },
                "major": {
                    "type": "integer",
                    "maximum": 65535,
                    "minimum": 0
                },
                "minor": {
                    "type": "integer",
                    "maximum": 65535,
                    "minimum": 0
                },
                "txPower": {
                    "description": "RSSI at 1m",
                    "type": "integer"
                },
                "uuid": {
                    "description": "iBeacon UUID",
                    "type": "string"
                },
                "x": {
                    "description": "meters",
                    "type": "number"
                },
                "y": {
                    "description": "meters",
                    "type": "number"
                }
            }
        },
        "trackerHub_backend_internal_models.WebUIConfig": {
            "type": "object",
            "required": [
                "beacons",
                "settings"
            ],
            "properties": {
                "beacons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/trackerHub_backend_internal_models.WebUIBeaconConfig"
                    }
                },
                "map": {
                    "description": "Map can be optional initially",
                    "allOf": [
                        {
                            "$ref": "#/definitions/trackerHub_backend_internal_models.WebUIMapInfo"
                        }
                    ]
                },
                "settings": {
                    "$ref": "#/definitions/trackerHub_backend_internal_models.WebUISettings"
                }
            }
        },
        "trackerHub_backend_internal_models.WebUIMapEntity": {
            "type": "object",
            "required": [
                "points",
                "type"
            ],
            "properties": {
                "closed": {
                    "type": "boolean"
                },
                "fillColor": {
                    "description": "If supporting filled shapes",
                    "type": "string"
                },
                "lineWidth": {
                    "description": "in pixels",
                    "type": "number"
                },
                "points": {
                    "description": "List of [x, y] coordinates",
                    "type": "array",
                    "minItems": 2,
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "number",
                            "format": "float64"
                        }
                    }
                },
                "strokeColor": {
                    "description": "e.g., '#333333'",
                    "type": "string"
                },
                "type": {
                    "description": "e.g., 'polyline'",
                    "type": "string",
                    "enum": [
                        "polyline"
                    ]
                }
            }
        },
        "trackerHub_backend_internal_models.WebUIMapInfo": {
            "type": "object",
            "required": [
                "height",
                "width"
            ],
            "properties": {
                "entities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/trackerHub_backend_internal_models.WebUIMapEntity"
                    }
                },
                "height": {
                    "description": "in meters",
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "width": {
                    "description": "in meters",
                    "type": "number"
                }
            }
        },
        "trackerHub_backend_internal_models.WebUISettings": {
            "type": "object",
            "required": [
                "signalPropagationFactor"
            ],
            "properties": {
                "signalPropagationFactor": {
                    "description": "Path loss exponent 'n'",
                    "type": "number",
                    "maximum": 6,
                    "minimum": 1
                }
            }
        }
    }
}